+++ packages/ai/CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## [0.13.0](https://github.com/axiomhq/ai/compare/axiom-v0.12.0...axiom-v0.13.0) (2025-08-20)
+
+
+### Features
+
+* **evals:** user custom names for cases ([#60](https://github.com/axiomhq/ai/issues/60)) ([e868ad6](https://github.com/axiomhq/ai/commit/e868ad6156f36ebca1e9ffd1aa69f58e47772ec8))
+
+## [0.12.0](https://github.com/axiomhq/ai/compare/axiom-v0.11.2...axiom-v0.12.0) (2025-08-17)
+
+
+### Features
+
+* better stream handling ([#57](https://github.com/axiomhq/ai/issues/57)) ([68c5a4d](https://github.com/axiomhq/ai/commit/68c5a4df3cbc1c30ad858817e8592d8a275c4014))
+
+## [0.11.2](https://github.com/axiomhq/ai/compare/axiom-v0.11.1...axiom-v0.11.2) (2025-08-15)
+
+
+### Features
+
+* **evals:** send score errors and tighten the scorer type ([#55](https://github.com/axiomhq/ai/issues/55)) ([85044f5](https://github.com/axiomhq/ai/commit/85044f529eefe4a767cbb82780dcc3ed869e4723))
+
+## [0.11.1](https://github.com/axiomhq/ai/compare/axiom-v0.11.0...axiom-v0.11.1) (2025-08-15)
+
+
+### Features
+
+* **evals:** attach user info to eval span ([#53](https://github.com/axiomhq/ai/issues/53)) ([449674e](https://github.com/axiomhq/ai/commit/449674e3a89ca8f0672b0cdbf0e40ab1f1cd1c2a))
+
+## [0.11.0](https://github.com/axiomhq/ai/compare/axiom-v0.10.0...axiom-v0.11.0) (2025-08-14)
+
+
+### ⚠ BREAKING CHANGES
+
+* change cli run command to eval [#49]
+* print summary table after eval runs [#49]
+
 ## [0.10.0](https://github.com/axiomhq/ai/compare/axiom-v0.8.0...axiom-v0.10.0) (2025-08-11)